requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==21.0.1
werkzeug==3.0.1
gunicorn==21.2.0
httpx==0.27.2
supabase==2.3.4